Move wait conditions and status into modules.
