+++ acceptancetests/jujupy/__init__.py
@@ -1,3 +1,18 @@
+# This file is part of JujuPy, a library for driving the Juju CLI.
+# Copyright 2013-2017 Canonical Ltd.
+#
+# This program is free software: you can redistribute it and/or modify it
+# under the terms of the Lesser GNU General Public License version 3, as
+# published by the Free Software Foundation.
+#
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranties of MERCHANTABILITY,
+# SATISFACTORY QUALITY, or FITNESS FOR A PARTICULAR PURPOSE.  See the Lesser
+# GNU General Public License for more details.
+#
+# You should have received a copy of the Lesser GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
 from jujupy.exceptions import (
     AgentsNotStarted,
     AuthNotAccepted,
@@ -8,7 +23,6 @@ from jujupy.exceptions import (
     TypeNotAccepted,
     )
 from jujupy.client import (
-    ConditionList,
     client_from_config,
     client_for_existing,
     get_cache_path,
@@ -25,7 +39,6 @@ from jujupy.client import (
     Machine,
     ModelClient,
     parse_new_state_server_from_error,
-    Status,
     temp_bootstrap_env,
     )
 from jujupy.configuration import (
@@ -38,7 +51,12 @@ from jujupy.fake import (
     FakeControllerState,
     fake_juju_client,
     )
-
+from jujupy.status import (
+    Status,
+    )
+from jujupy.wait_condition import (
+    ConditionList,
+    )
 
 __all__ = [
     'AgentsNotStarted',
